a bunch more

+++ src/controller/ButtonMapping.cpp
@@ -1,6 +1,6 @@
 #include "ButtonMapping.h"
 
 namespace LUS {
-ButtonMapping::ButtonMapping(int32_t bitmask) : mBitmask(bitmask) {
+ButtonMapping::ButtonMapping(uint16_t bitmask) : mBitmask(bitmask) {
 }
 } // namespace LUS
