+++ src/controller/ButtonMapping.h
@@ -3,12 +3,12 @@
 namespace LUS {
 class ButtonMapping {
   public:
-    ButtonMapping(int32_t bitmask);
+    ButtonMapping(uint16_t bitmask);
     ~ButtonMapping();
 
-    virtual void UpdatePad(int32_t& padButtons) = 0;
+    virtual void UpdatePad(uint16_t& padButtons) = 0;
 
   protected:
-    int32_t mBitmask;
+    uint16_t mBitmask;
 };
 } // namespace LUS
