+++ src/controller/DummyButtonMapping.h
@@ -4,6 +4,6 @@ namespace LUS {
 class DummyButtonMapping final : public ButtonMapping {
   public:
     using ButtonMapping::ButtonMapping;
-    void UpdatePad(int32_t& padButtons) override;
+    void UpdatePad(uint16_t& padButtons) override;
 };
 } // namespace LUS
