+++ src/controller/SDLAxisToButtonMapping.h
@@ -6,8 +6,8 @@ enum AxisDirection { NEGATIVE = 0, POSITIVE = 1};
 
 class SDLAxisToButtonMapping final : public ButtonMapping, public SDLMapping {
   public:
-    SDLAxisToButtonMapping(int32_t bitmask, int32_t sdlControllerIndex, int32_t sdlControllerAxis, int32_t axisDirection);
-    void UpdatePad(int32_t& padButtons) override;
+    SDLAxisToButtonMapping(uint16_t bitmask, int32_t sdlControllerIndex, int32_t sdlControllerAxis, int32_t axisDirection);
+    void UpdatePad(uint16_t& padButtons) override;
 
   private:
     SDL_GameControllerAxis mControllerAxis;
